+++ crates/runtime/src/vmcontext.rs
@@ -203,7 +203,7 @@ pub struct VMMemoryDefinition {
     pub base: *mut u8,
 
     /// The current logical size of this linear memory in bytes.
-    pub current_length: usize,
+    pub current_length: u32,
 }
 
 #[cfg(test)]
